Use fixed-sized types for return values.
sv_flags is defined as U32, so that should be the return type of SvROK()
as well.

SvTYPE() returns an enum value, which is problematic since enum sizes
are not standardized and compilers have a lot of freedom there. In this
particular case, IV should probably be always enough to contain any
possible SVt_* value.