Make the stack pointer more explicit.
917fdb7 tried to unify naming of stack type between %Size and arguments
to functions, assuming that stack_t is a structure type. Until now, it
was actually a pointer type, so the size in %Size hash was reported
incorrectly.

This commit changes stack_t to be the structure type and makes all
pointers to it explicit to avoid possibility of similar breakage in the
future.